requirements.txt
PySide6>=6.6
google-api-python-client>=2.0
google-auth>=2.0
requests>=2.28
beautifulsoup4>=4.12
networkx>=3.0
pandas>=2.0
lxml>=4.9
